test(core): moving tests close to their usage
Rather than 1 tests folders we have many.
