+++ packages/desktop-core/src/main/configuration/loaders/__tests__/directoryConfigurationLoader.test.ts
@@ -1,7 +1,7 @@
 import { mock } from "jest-mock-extended";
 
-import { DirectoryConfigurationLoader } from "../../../src/main/configuration/loaders/directoryConfigurationLoader";
-import { IConfigurationLoader } from "../../../src/main/configuration/loaders/iConfigurationLoader";
+import { DirectoryConfigurationLoader } from "../directoryConfigurationLoader";
+import { IConfigurationLoader } from "../iConfigurationLoader";
 
 describe("canLoad", () => {
