+++ packages/desktop-core/src/main/configuration/loaders/__tests__/localFileConfigurationLoader.test.ts
@@ -1,7 +1,7 @@
 import { mock } from "jest-mock-extended";
 
-import { LocalFileConfigurationLoader } from "../../../src/main/configuration/loaders/localFileConfigurationLoader";
-import { IConfigurationParser } from "../../../src/main/configuration/parsers/iConfigurationParser";
+import { IConfigurationParser } from "../../parsers/iConfigurationParser";
+import { LocalFileConfigurationLoader } from "../localFileConfigurationLoader";
 
 describe("canLoad", () => {
