+++ packages/desktop-core/src/main/configuration/loaders/__tests__/remoteFileConfigurationLoader.test.ts
@@ -1,7 +1,7 @@
 import { mock } from "jest-mock-extended";
 
-import { RemoteFileConfigurationLoader } from "../../../src/main/configuration/loaders/remoteFileConfigurationLoader";
-import { IConfigurationParser } from "../../../src/main/configuration/parsers/iConfigurationParser";
+import { IConfigurationParser } from "../../parsers/iConfigurationParser";
+import { RemoteFileConfigurationLoader } from "../remoteFileConfigurationLoader";
 
 describe("canLoad", () => {
