+++ packages/desktop-core/src/main/configuration/parsers/__tests__/jsonConfigurationParser.test.ts
@@ -1,4 +1,4 @@
-import { JsonConfigurationParser } from "../../../src/main/configuration/parsers/jsonConfigurationParser";
+import { JsonConfigurationParser } from "../jsonConfigurationParser";
 
 describe("parse", () => {
