+++ packages/desktop-core/src/main/registry/__tests__/priorityConfigurationRegistryService.test.ts
@@ -1,9 +1,9 @@
 import { mock } from "jest-mock-extended";
 
-import { ConfigurationKind } from "../../src/main/configuration/configurationKind";
-import { IConfiguration } from "../../src/main/configuration/iConfiguration";
-import { IRegistryService } from "../../src/main/registry/iRegistryService";
-import { PriorityConfigurationRegistryService } from "../../src/main/registry/priorityConfigurationRegistryService";
+import { ConfigurationKind } from "../../configuration/configurationKind";
+import { IConfiguration } from "../../configuration/iConfiguration";
+import { IRegistryService } from "../iRegistryService";
+import { PriorityConfigurationRegistryService } from "../priorityConfigurationRegistryService";
 
 describe("getRegistry", () => {
